+++ doc/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog {#Changelog}
 
+# git master
+
+* [285](https://github.com/Eyescale/Lunchbox/pull/285):
+  Allow custom alloc/dealloc functors in Buffer.
+
 # Release 1.16 (23-05-2017)
 
 * [310](https://github.com/Eyescale/Lunchbox/pull/310):